Graphics.Farbe

Description

A OpenGL ES 2 rendering library.

GLES2 is ancient, but good enough for most things~

A sample program:

import Graphics.Farbe
import Data.Function (fix)

main :: IO ()
main = runFarbeT "Window title" (InWindow (1000,800)) $ renderColorful

renderColorful :: (MonadWindow m, Farbe m) => m ()
renderColorful = do
  va <- newVArray frame
  fix $ \loop -> processEvents $ \es -> do
    r <- rotationFromMouse33
    runShader colorful r [va]
    loop

colorful
  :: Mat V3 V3 (Expr V Float)               -- rotation matrix parameter
  -> (V3 (Expr V Float))                    -- vertex array attribute parameter
  -> (V4 (Expr V Float), V4 (Expr F Float)) -- shader result
colorful r v = let
  v' = r **| v
  n' = transferFrag v
  in (up 1 v', up 1 n' * 0.5 + 0.2)

Farbe

runFarbeT :: (MonadIO m, Typeable m) => String -> Display -> FarbeT (WindowT m) a -> m a

The environment to do draw operations. It spawns a window with the render context.

data Display

Constructors

Fullscreen

InWindow (Int, Int)

InWindowAt (Int, Int) (Int, Int)

Event handling

processEvents :: (MonadWindow m, Farbe m) => ([(Event, EventContext)] -> m ()) -> m ()

processEvents Sends window events to a provided function. The function is always called, as long as the program isn't asked to quit. This function also controls the render pipeline - swapping buffers, clear screen, etc.

data Event

Constructors

EventKey Key KeyState String
The Key is mapped to US keyboard layout and String is the localized name of the key.

EventTyping Char
Localized character. Use these for text fields.

EventMouseKey (Float, Float) MouseButton KeyState
Coordinate origin (0,0) is left top corner.

EventMouseMove (Float, Float)

EventMouseKeyLocked MouseButton KeyState
See locked state.

EventMouseMoveLocked (Float, Float)

EventScroll Float Float

EventFocus Bool

EventMouseInOutWindow Bool

EventDrop [String]

EventClose

data Key

Part of the Keyboard Input system.

Constructors

Key'Unknown

Key'Space

Key'Apostrophe

Key'Comma

Key'Minus

Key'Period

Key'Slash

Key'0

Key'1

Key'2

Key'3

Key'4

Key'5

Key'6

Key'7

Key'8

Key'9

Key'Semicolon

Key'Equal

Key'A

Key'B

Key'C

Key'D

Key'E

Key'F

Key'G

Key'H

Key'I

Key'J

Key'K

Key'L

Key'M

Key'N

Key'O

Key'P

Key'Q

Key'R

Key'S

Key'T

Key'U

Key'V

Key'W

Key'X

Key'Y

Key'Z

Key'LeftBracket

Key'Backslash

Key'RightBracket

Key'GraveAccent

Key'World1

Key'World2

Key'Escape

Key'Enter

Key'Tab

Key'Backspace

Key'Insert

Key'Delete

Key'Right

Key'Left

Key'Down

Key'Up

Key'PageUp

Key'PageDown

Key'Home

Key'End

Key'CapsLock

Key'ScrollLock

Key'NumLock

Key'PrintScreen

Key'Pause

Key'F1

Key'F2

Key'F3

Key'F4

Key'F5

Key'F6

Key'F7

Key'F8

Key'F9

Key'F10

Key'F11

Key'F12

Key'F13

Key'F14

Key'F15

Key'F16

Key'F17

Key'F18

Key'F19

Key'F20

Key'F21

Key'F22

Key'F23

Key'F24

Key'F25

Key'Pad0

Key'Pad1

Key'Pad2

Key'Pad3

Key'Pad4

Key'Pad5

Key'Pad6

Key'Pad7

Key'Pad8

Key'Pad9

Key'PadDecimal

Key'PadDivide

Key'PadMultiply

Key'PadSubtract

Key'PadAdd

Key'PadEnter

Key'PadEqual

Key'LeftShift

Key'LeftControl

Key'LeftAlt

Key'LeftSuper

Key'RightShift

Key'RightControl

Key'RightAlt

Key'RightSuper

Key'Menu

data KeyState

Constructors

Down

Up

anyMouseClick :: Monad m => [(Event, EventContext)] -> m () -> m ()

rotationFromMouse33 :: MonadWindow m => m (Mat V3 V3 Float)

viewMat :: MonadWindow m => m (Mat V4 V4 Float)

Shader definition

Example:

foo :: Farbe m => V3 (V3 Float) -> [VArray (V3 Float)] -> m Bool
foo = runShader colorful

colorful :: Mat V3 V3 (Expr V Float) -> (V3 (Expr V Float)) -> (V4 (Expr V Float), V4 (Expr F Float))
colorful = undefined

A shader function is defined by

a number of Uniform parameters

one Attribute parameter

a result

Uniforms are defined by the Use class and convert values as follows:

Real space	Shader space

Float	(Expr e Float)

Int32	(Expr e Int32)

Bool	(Expr e Bool)

Texture	(Expr e Texture)

(V2 Float)	(V2 (Expr e Float))

(V2 Int32)	(V2 (Expr e Int32))

(V2 Bool)	(V2 (Expr e Bool))

(V3 Float)	(V3 (Expr e Float))

(V3 Int32)	(V3 (Expr e Int32))

(V3 Bool)	(V3 (Expr e Bool))

(V4 Float)	(V4 (Expr e Float))

(V4 Int32)	(V4 (Expr e Int32))

(V4 Bool)	(V4 (Expr e Bool))

(V2 (V2 Float))	(V2 (V2 (Expr e Float)))

(V3 (V3 Float))	(V3 (V3 (Expr e Float)))

(V4 (V4 Float))	(V4 (V4 (Expr e Float)))

Attributes are defined by Attribute in a scheme similar to Uniforms.

The result of a shader is a 2-tuple of:

(V4 (Expr V Float)) determines the position of a vertex of a triangle in a render of coordinates -1 to 1 on x, y and z . w as homogeneous coordinate.

(V4 (Expr F Float)) determines the color in rgb for every rendered pixel.

type SResult = (V4 (Expr V Float), V4 (Expr F Float))

runShader :: forall h f (m :: Type -> Type) g. (UseFun h f, MonadIO m, Shader m f g, JoinF m g, HandShdr m, Typeable g) => h -> g

Run shader function with accessible parameters.

runShaderV :: forall (m :: Type -> Type) f g. (MonadIO m, Shader m f g, JoinF m g, HandShdr m, Typeable g) => f -> g

Run shader function with its parameters as Var. Var are accessible through use within all shader spaces. runShaderV runs with functional dependencies and can be used to look up the generated function definition.

Graphics.Farbe.Vec is the default math module for Farbe.

module Graphics.Farbe.Vec

data Expr (e :: k) (a :: k1)

Expression type to define shader definitions.

data V

data F

fragCoord :: V4 (Expr F Float)

position coordinate in Fragment shader.

windowDim :: MonadWindow m => m (V2 Float)

napier :: Fractional a => a

e

ln :: Floating a => a -> a

modf :: forall {k} (e :: k). Expr e Float -> Expr e Float -> Expr e Float

modulo on float.

equot :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

erem :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

ediv :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

emod :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

transferFrag :: Transfer a b => a -> b

texture :: forall {k} (e :: k). Expr e Texture -> V2 (Expr e Float) -> V4 (Expr e Float)

Access values of a texture based on given coordinates.

Vertex array

data VArray (a :: k)

Vertex array, saving model data for shaders on the GPU.

newVArray :: (MonadIO m, HandVBO m, Storable a, Foldable f, Attribute a b) => f a -> m (VArray a)

class Storable a => Attribute a b | a -> b, b -> a

Minimal complete definition

setAttribute

frame :: [V3 Float]

A plane covering the entire view with two triangles covering the visible front.

floorFrame :: [V3 Float]

Textures on shaders

data Texture

loadImage :: (MonadIO m, Farbe m) => String -> m Texture

loadImagePixellated :: (MonadIO m, Farbe m) => String -> m Texture

Render control

These functions are for combining multiple renders, using mostly stencils internally.

WIP likely broken functions.

drawOver :: MonadIO m => m a1 -> m a2 -> m ()

drawTexture :: (Monad m, Farbe m, MonadWindow m) => m (m () -> m Texture)

drawDepth :: (Monad m, Farbe m, MonadWindow m) => m (m () -> m Texture)

drawInto :: MonadIO m => m a1 -> m a2 -> m ()

Configuration options

modifyConfig :: Farbe m => (Config -> Config) -> m ()

data Config

Constructors

Config

Fields

workTime :: Double

backgroundColor :: V4 Float

Miscellaneous

class Use (a :: k) r | r -> a

Minimal complete definition

use

data FarbeT (m :: Type -> Type) a

class (MonadIO m, Typeable m) => Farbe (m :: Type -> Type)

Minimal complete definition

stateFarbe

runFarbeT' :: (MonadIO m, Typeable m) => FarbeT m a -> m a

class MonadIO m => MonadWindow (m :: Type -> Type)

Minimal complete definition

windowState

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3