Graphics.Farbe.Expr

Documentation

data Expr (e :: k) (a :: k1)

Expression type to define shader definitions.

Constructors

Expr

Fields

unExpr :: ExprI

data ExprI

Constructors

ExprI

Fields

fnName :: String

rtype :: TypeS

fnAst :: [ExprI]

exprSetup :: Register

data Register

Constructors

RegisterNone

RegisterVarying

RegisterUniform

RegisterVertex

data F

data V

type SResult = (V4 (Expr V Float), V4 (Expr F Float))

liftExpr :: forall {k} a (e :: k). GLtype a => String -> [ExprI] -> Expr e a

liftE0 :: forall {k} a (e :: k). GLtype a => String -> Expr e a

liftE1 :: forall {k1} {k2} a2 (e :: k1) (a1 :: k2). GLtype a2 => String -> Expr e a1 -> Expr e a2

liftE2 :: forall {k1} {k2} {k3} a3 (e :: k1) (a1 :: k2) (a2 :: k3). GLtype a3 => String -> Expr e a1 -> Expr e a2 -> Expr e a3

liftE3 :: forall {k1} {k2} {k3} {k4} a4 (e :: k1) (a1 :: k2) (a2 :: k3) (a3 :: k4). GLtype a4 => String -> Expr e a1 -> Expr e a2 -> Expr e a3 -> Expr e a4

crawl :: (ExprI -> a) -> ExprI -> [a]

mapExpr :: Monad m => (ExprI -> m ExprI) -> ExprI -> m ExprI

napier :: Fractional a => a

e

ln :: Floating a => a -> a

modf :: forall {k} (e :: k). Expr e Float -> Expr e Float -> Expr e Float

modulo on float.

log2 :: forall {k} (e :: k). Expr e Float -> Expr e Float -> Expr e Float

modulo on float.

efloor :: forall {k} (e :: k). Expr e Float -> Expr e Float

equot :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

erem :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

ediv :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

emod :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

Expression space equivalent.

vecParts :: forall {k} a v (e :: k). (GLtype a, Vector v) => Expr e (v a) -> v (Expr e a)

matParts :: forall {k} a v (e :: k). (GLtype a, GLtype (v a), Vector v) => Expr e (v (v a)) -> v (v (Expr e a))

exprVec :: forall {k} (e :: k) a v. (GLtype a, Vector v, GLtype (v a)) => v (Expr e a) -> Expr e (v a)

exprMat :: forall {k} a (e :: k) v. (GLtype a, Vector v, GLtype (v a), GLtype (v (v a))) => v (v (Expr e a)) -> Expr e (v (v a))

arrV :: forall {k} a v (e :: k). (GLtype a, Vector v) => Expr e (v a) -> Expr e Int32 -> Expr e a

literal :: forall {k} b a (e :: k). (Show b, GLtype a) => b -> Expr e a

fragCoord :: V4 (Expr F Float)

position coordinate in Fragment shader.

texture :: forall {k} (e :: k). Expr e Texture -> V2 (Expr e Float) -> V4 (Expr e Float)

Access values of a texture based on given coordinates.

texture' :: forall {k} (e :: k). Expr e Texture -> V2 (Expr e Float) -> V4 (Expr e Float)

Original unflipped texture variables.

arr :: forall {k} a (e :: k) (s :: Nat). GLtype a => Expr e (Arr s a) -> Int32 -> Expr e a

arr' :: forall {k} a (e :: k) (s :: Nat). GLtype a => Expr e (Arr s a) -> Expr e Int32 -> Expr e a

arr' is ignoring constant expression requirement. May not work with some implementations.

if' :: forall {k} a (e :: k). GLtype a => Expr e Bool -> Expr e a -> Expr e a -> Expr e a

Expr space if.

class Transfer a b | a -> b, b -> a where

Transmitting values from vertex to fragment shader. If a value is transferred from the rendered vertex array directly, they are interpolated to their coordinate in fragment space.

Methods

transferFrag :: a -> b