Graphics.Farbe.Shader

Documentation

data ShdrState

Constructors

ShdrState

Fields

shdrMap :: IntMap Dynamic

shdrCompState :: IntMap [String]

initShdrState :: ShdrState

class Monad m => HandShdr (m :: Type -> Type) where

Methods

stateShdr :: (ShdrState -> (a, ShdrState)) -> m a

getShdr :: HandShdr m => m ShdrState

setShdr :: HandShdr m => ShdrState -> m ()

stateShdrMap :: HandShdr m => (IntMap Dynamic -> (a, IntMap Dynamic)) -> m a

getShdrMap :: HandShdr m => m (IntMap Dynamic)

modifyShdrMap :: HandShdr m => (IntMap Dynamic -> IntMap Dynamic) -> m (IntMap Dynamic)

class Shader (m :: Type -> Type) f g | m f -> g, g -> f, g -> m where

Methods

mkShader :: f -> ShaderEnvT m g

idShader :: f -> m Int

setUniform :: forall (m :: Type -> Type) u g f. (MonadIO m, Uniform u, HandTex m, AppliableF m g, Shader m f g) => Int -> (Var u -> f) -> ShaderEnvT m (u -> g)

setIdShader :: (MonadIO m, Shader m f g, Uniform a) => Int -> (Var a -> f) -> m Int

isProgramCompiled :: MonadIO m => ShaderId -> m Bool

shaderCompileProgress :: f -> m [(String, Bool)]

runShaderV' :: (MonadIO m, Shader m f g, HandShdr m, Typeable g) => f -> m g

runShaderV :: forall (m :: Type -> Type) f g. (MonadIO m, Shader m f g, JoinF m g, HandShdr m, Typeable g) => f -> g

Run shader function with its parameters as Var. Var are accessible through use within all shader spaces. runShaderV runs with functional dependencies and can be used to look up the generated function definition.

runShader :: forall h f (m :: Type -> Type) g. (UseFun h f, MonadIO m, Shader m f g, JoinF m g, HandShdr m, Typeable g) => h -> g

Run shader function with accessible parameters.

compileShader :: (MonadIO m, Shader m f g, HandShdr m, Typeable g) => f -> m g

fmap3 :: (Functor f1, Functor f2, Functor f3) => (a -> b) -> f1 (f2 (f3 a)) -> f1 (f2 (f3 b))

optimizeExpressions :: ShaderEnv m => m ()

optVectors :: ExprI -> ExprI

same :: Eq a => [a] -> Bool

handleTransfers :: ShaderEnv m => m ()

collectHeaders :: ShaderEnv m => m ()

addInputHeader :: ShaderEnv m => ShaderType -> ExprI -> m ()

addVaryingOutputHeader :: ShaderEnv m => String -> ShaderType -> ExprI -> m ()

compileSubShader :: (MonadIO m, ShaderEnv m) => ShaderType -> m ()

toCExpr :: ExprI -> String

class JoinF (m :: Type -> Type) f where

Methods

joinF :: m f -> f

class AppliableF (m :: k -> Type) f | f -> m where

Methods

applyF :: forall (a :: k). f -> m a -> f

(.:) :: (b -> c) -> (a1 -> a2 -> b) -> a1 -> a2 -> c

data ShaderType

Constructors

Vertex

Fragment

data ShaderData

Constructors

ShaderData

Fields

counter :: Int

shaderId :: ShaderId

headers :: Set (ShaderType, Header)

exprs :: [(ShaderType, (String, ExprI))]

preRender :: IO ()

type Header = String

emptyShaderData :: MonadIO m => m ShaderData

class Monad m => ShaderEnv (m :: Type -> Type) where

Methods

stateShader :: (ShaderData -> (a, ShaderData)) -> m a

modifyShader :: ShaderEnv m => (ShaderData -> ShaderData) -> m ()

getsShader :: ShaderEnv m => (ShaderData -> r) -> m r

getShader :: ShaderEnv m => m ShaderData

putShader :: ShaderEnv m => ShaderData -> m ()

newtype ShaderEnvT (m :: Type -> Type) a

Constructors

ShaderEnvT

Fields

unShaderEnvT :: StateT ShaderData m a

runShaderEnvT :: MonadIO m => ShaderEnvT m a -> m (a, ShaderData)

evalShaderEnvT :: MonadIO m => ShaderEnvT m a -> m a

addExpr :: ShaderEnv m => ShaderType -> String -> ExprI -> m ()