Graphics.Farbe.Vec

Description

A basic math module for vectors and matrices.

Vector types

data V1 a

Constructors

V1 a

data V2 a

Constructors

V2 a a

data V3 a

Constructors

V3 a a a

data V4 a

Constructors

V4 a a a a

class (Applicative v, Foldable v, Traversable v, FromList v) => Vector (v :: Type -> Type) where

Unifying vector class.

Methods

vsize :: Num n => v a -> n

Vector operations

(*|) :: (Vector v, Num (v a)) => a -> v a -> v a

Scalar multiplication.

(*||) :: (Vector x, Vector y, Num (x (y a))) => a -> Mat x y a -> Mat x y a

Scalar multiplication using a matrix.

vdot :: (Vector v, Num (v a), Num a) => v a -> v a -> a

Dot product.

vlength :: (Vector v, Num (v a), Floating a) => v a -> a

Vector length.

vdistance :: (Vector v, Num (v a), Floating a) => v a -> v a -> a

Distance between two vector coordinates.

vnormal :: (Vector v, Fractional (v a), Floating a) => v a -> v a

Normalize vectors. Reduce a vector's length to 1, by retaining orientation.

vcross :: Num a => V3 a -> V3 a -> V3 a

Cross product.

rel :: Num v => v -> (v -> v) -> v -> v

The first parameter is substracted and then added back afterwards.

relm :: Fractional v => v -> (v -> v) -> v -> v

Same as rel for multiplication.

line :: (Vector v, Num (v a)) => v a -> v a -> a -> v a

Calculates a point on a line between its two parameters. The point is based on the third parameter, which is interpolated by values between 0 and 1 .

curve :: (Vector v, Num (v a)) => v a -> v a -> v a -> a -> v a

Bezier curve with one control point.

Matrix operations

type Mat (vx :: k1 -> k) (vy :: k2 -> k1) (a :: k2) = vx (vy a)

Matrix type.

mtranspose :: (Vector x, Vector y) => Mat x y a -> Mat y x a

Matrix transpose.

vtranspose :: Vector v => v a -> Mat v V1 a

Tranpose vertex to matrix.

vtranspose' :: Vector v => Mat v V1 a -> v a

Tranpose matrix to vertex.

mult :: (Vector m, Vector n, Vector h, Num (m a), Num (n a), Num (h a), Num a) => Mat m n a -> Mat n h a -> Mat m h a

Matrix multiplication.

(****) :: (Vector m, Vector n, Vector h, Num (m a), Num (n a), Num (h a), Num a) => Mat m n a -> Mat n h a -> Mat m h a

Matrix multiplication, alias.

multv :: (Vector m, Vector n, Num (m a), Num (n a), Num a) => Mat m n a -> n a -> m a

Matrix multiplication with vector.

(**|) :: (Vector m, Vector n, Num (m a), Num (n a), Num a) => Mat m n a -> n a -> m a

Matrix multiplication with vector, alias.

toList2 :: (Foldable f, Foldable g) => f (g a) -> [a]

Rotation

rotate2D :: Floating a => a -> V2 a -> V2 a

Rotate in 2D space.

rotationMatrix2D :: Floating a => a -> Mat V2 V2 a

Define a multiplication matrix for 2D space.

rotate :: Floating a => a -> a -> a -> V3 a -> V3 a

Rotate in 3D space.

roll :: Floating a => a -> Mat V3 V3 a

pitch :: Floating a => a -> Mat V3 V3 a

yaw :: Floating a => a -> Mat V3 V3 a

rotationMatrix :: Floating a => a -> a -> a -> Mat V3 V3 a

Define a multiplication matrix for 3D space.

Perspective

perspective

Arguments

:: Floating a

=> a
FOV (y direction, in radians)

-> a
Aspect ratio

-> a
Near plane

-> a
Far plane

-> Mat V4 V4 a

Multiplication matrix for orthogonal to perspective projection.

Utility classes

class FromList (t :: Type -> Type) where

Methods

fromListFill :: a -> [a] -> t a

fromList :: (FromList t, Num a) => [a] -> t a

class ToTuple a b where

Methods

toTuple :: a -> b

class FromTuple a b where

Methods

fromTuple :: a -> b

class GetX (v :: Type -> Type) where

Methods

getx :: v a -> a

class GetY (v :: Type -> Type) where

Methods

gety :: v a -> a

class GetZ (v :: Type -> Type) where

Methods

getz :: v a -> a

class GetW (v :: Type -> Type) where

Methods

getw :: v a -> a

class Fit x a b | a -> x, b -> x where

Class to coerce to fitting size.

Methods

fit :: x -> a -> b

fit0 :: (Num a, Fit a (v a) (v2 a)) => v a -> v2 a

fit1 :: (Num a, Fit a (v a) (v2 a)) => v a -> v2 a

class SizeUp x a b | a -> x, b -> x where

Class to size up vector to asked size.

Methods

up :: x -> a -> b

class SizeDown a b where

Class to size down vector to asked size.

Methods

down :: a -> b

Miscellaneous

subSizeOf :: forall g a n. (Storable a, Num n) => g a -> n

sizeOf variant for type parameter.

itoi :: (Integral a, Num c) => a -> c

for :: [a] -> (a -> b) -> [b]

map with flipped parameters.