Graphics.Farbe.VertexArray

Documentation

newVArray :: (MonadIO m, HandVBO m, Storable a, Foldable f, Attribute a b) => f a -> m (VArray a)

drawArrays :: (MonadIO m, Storable a) => [VArray a] -> m ()

data VArray (a :: k)

Vertex array, saving model data for shaders on the GPU.

class Monad m => HandVBO (m :: Type -> Type) where

Minimal complete definition

stateVBO, delayVBO

Methods

stateVBO :: (VBOState -> (a, VBOState)) -> m a

stateVBO' :: (VBOState -> VBOState) -> m ()

delayVBO :: m ((VBOState -> VBOState) -> IO ())

data VBOState

initHandVBOState :: MonadIO m => GLintptr -> m VBOState

simple arrays

frame :: [V3 Float]

A plane covering the entire view with two triangles covering the visible front.

floorFrame :: [V3 Float]