Graphics.Farbe.Window

Description

A basic window interface for running a single window. Has to run on main thread and is not thread-safe.

This module provides only a fraction of GLFW's abilities. If you are missing functionality, you should copy/fork this module and adjust it to your needs.

Window creation

runWindowT :: MonadIO m => String -> Display -> WindowT m a -> m a

Creates a fullscreen window, runs your action and terminates window after.

data Display

Constructors

Fullscreen

InWindow (Int, Int)

InWindowAt (Int, Int) (Int, Int)

newtype WindowT (m :: Type -> Type) a

Constructors

WindowT

Fields

runw :: ReaderT WindowState m a

class MonadIO m => MonadWindow (m :: Type -> Type)

Minimal complete definition

windowState

swapBuffers :: MonadWindow m => m ()

Finish render and display it on screen.

Event processing

processEvents :: MonadWindow m => m [(Event, EventContext)]

Process and fetch events.

shouldWindowClose :: MonadWindow m => m Bool

Process and fetch events. Delivers Events to a function. The function is run, unless window is signaled to be closed. ~ processEvents :: MonadWindow m => ([(Event, EventContext)] -> m ()) -> m () ~ processEvents f = do ~ es <- processEvents' ~ w <- glfwWindow ~ b <- liftIO $ W.windowShouldClose w ~ when (not $ b || elem EventClose (map fst es)) $ f es

data Event

Constructors

EventKey Key KeyState String
The Key is mapped to US keyboard layout and String is the localized name of the key.

EventTyping Char
Localized character. Use these for text fields.

EventMouseKey (Float, Float) MouseButton KeyState
Coordinate origin (0,0) is left top corner.

EventMouseMove (Float, Float)

EventMouseKeyLocked MouseButton KeyState
See locked state.

EventMouseMoveLocked (Float, Float)

EventScroll Float Float

EventFocus Bool

EventMouseInOutWindow Bool

EventDrop [String]

EventClose

type EventContext = Set (Either MouseButton Key)

Cursor mode

setCursorMode :: MonadWindow m => CursorMode -> m ()

Set cursor input mode. When locked, the mouse cursor is hidden and can't leave the window. Mouse move and clicks will give locked variants of events instead. Unlike normal mouse position info from events, locked mouse move gives relative position since last event processing.

getCursorMode :: MonadWindow m => m CursorMode

data CursorMode

Constructors

CursorNormal

CursorHidden

CursorLocked

data KeyState

Constructors

Down

Up

GLFW re-export

data Key

Part of the Keyboard Input system.

Constructors

Key'Unknown

Key'Space

Key'Apostrophe

Key'Comma

Key'Minus

Key'Period

Key'Slash

Key'0

Key'1

Key'2

Key'3

Key'4

Key'5

Key'6

Key'7

Key'8

Key'9

Key'Semicolon

Key'Equal

Key'A

Key'B

Key'C

Key'D

Key'E

Key'F

Key'G

Key'H

Key'I

Key'J

Key'K

Key'L

Key'M

Key'N

Key'O

Key'P

Key'Q

Key'R

Key'S

Key'T

Key'U

Key'V

Key'W

Key'X

Key'Y

Key'Z

Key'LeftBracket

Key'Backslash

Key'RightBracket

Key'GraveAccent

Key'World1

Key'World2

Key'Escape

Key'Enter

Key'Tab

Key'Backspace

Key'Insert

Key'Delete

Key'Right

Key'Left

Key'Down

Key'Up

Key'PageUp

Key'PageDown

Key'Home

Key'End

Key'CapsLock

Key'ScrollLock

Key'NumLock

Key'PrintScreen

Key'Pause

Key'F1

Key'F2

Key'F3

Key'F4

Key'F5

Key'F6

Key'F7

Key'F8

Key'F9

Key'F10

Key'F11

Key'F12

Key'F13

Key'F14

Key'F15

Key'F16

Key'F17

Key'F18

Key'F19

Key'F20

Key'F21

Key'F22

Key'F23

Key'F24

Key'F25

Key'Pad0

Key'Pad1

Key'Pad2

Key'Pad3

Key'Pad4

Key'Pad5

Key'Pad6

Key'Pad7

Key'Pad8

Key'Pad9

Key'PadDecimal

Key'PadDivide

Key'PadMultiply

Key'PadSubtract

Key'PadAdd

Key'PadEnter

Key'PadEqual

Key'LeftShift

Key'LeftControl

Key'LeftAlt

Key'LeftSuper

Key'RightShift

Key'RightControl

Key'RightAlt

Key'RightSuper

Key'Menu

data MouseButton

Part of the Mouse Input system.

Constructors

MouseButton'1

MouseButton'2

MouseButton'3

MouseButton'4

MouseButton'5

MouseButton'6

MouseButton'7

MouseButton'8

Error handling

data WindowErr

The functions of this module throw exceptions.

Constructors

ErrNoMonitor

ErrNoVideoMode

ErrWindowCreationFailed

ErrNoContext

ErrGLFWInitFailed

ErrApiUnavailable

ErrVersionUnavailable

ErrPlatformError

ErrElse String

Utility

getTime :: (MonadWindow m, Fractional a) => m a

Returns time in seconds since window launch.

windowSize :: (MonadIO m, MonadWindow m) => m (Int, Int)

glfwWindow :: MonadWindow m => m Window

Access to underlying GLFW window.