Graphics.Farbe

Description

This library abstracts away traps of OpenGL and provides its basics for rendering.

Documentation

runFarbeT :: MonadIO m => String -> Display -> FarbeT (WindowT m) a -> m a

The environment to do draw operations. It spawns a window with the render context.

data Display

Constructors

Fullscreen

InWindow (Int, Int)

InWindowAt (Int, Int) (Int, Int)

Event handling

processEvents :: (MonadWindow m, Farbe m) => ([(Event, EventContext)] -> m ()) -> m ()

Key function to get events and render to window (swap buffers)

data Event

Constructors

EventKey Key KeyState String
The Key is mapped to US keyboard layout and String is the localized name of the key.

EventTyping Char
Localized character. Use these for text fields.

EventMouseKey (Float, Float) MouseButton KeyState
Coordinate origin (0,0) is left top corner.

EventMouseMove (Float, Float)

EventMouseKeyLocked MouseButton KeyState
See locked state.

EventMouseMoveLocked (Float, Float)

EventScroll Float Float

EventFocus Bool

EventMouseInOutWindow Bool

EventDrop [String]

EventClose

data Key

Part of the Keyboard Input system.

Constructors

Key'Unknown

Key'Space

Key'Apostrophe

Key'Comma

Key'Minus

Key'Period

Key'Slash

Key'0

Key'1

Key'2

Key'3

Key'4

Key'5

Key'6

Key'7

Key'8

Key'9

Key'Semicolon

Key'Equal

Key'A

Key'B

Key'C

Key'D

Key'E

Key'F

Key'G

Key'H

Key'I

Key'J

Key'K

Key'L

Key'M

Key'N

Key'O

Key'P

Key'Q

Key'R

Key'S

Key'T

Key'U

Key'V

Key'W

Key'X

Key'Y

Key'Z

Key'LeftBracket

Key'Backslash

Key'RightBracket

Key'GraveAccent

Key'World1

Key'World2

Key'Escape

Key'Enter

Key'Tab

Key'Backspace

Key'Insert

Key'Delete

Key'Right

Key'Left

Key'Down

Key'Up

Key'PageUp

Key'PageDown

Key'Home

Key'End

Key'CapsLock

Key'ScrollLock

Key'NumLock

Key'PrintScreen

Key'Pause

Key'F1

Key'F2

Key'F3

Key'F4

Key'F5

Key'F6

Key'F7

Key'F8

Key'F9

Key'F10

Key'F11

Key'F12

Key'F13

Key'F14

Key'F15

Key'F16

Key'F17

Key'F18

Key'F19

Key'F20

Key'F21

Key'F22

Key'F23

Key'F24

Key'F25

Key'Pad0

Key'Pad1

Key'Pad2

Key'Pad3

Key'Pad4

Key'Pad5

Key'Pad6

Key'Pad7

Key'Pad8

Key'Pad9

Key'PadDecimal

Key'PadDivide

Key'PadMultiply

Key'PadSubtract

Key'PadAdd

Key'PadEnter

Key'PadEqual

Key'LeftShift

Key'LeftControl

Key'LeftAlt

Key'LeftSuper

Key'RightShift

Key'RightControl

Key'RightAlt

Key'RightSuper

Key'Menu

data KeyState

Constructors

Down

Up

Shader definition

shader :: (Farbe m, AttrType a b) => (b -> ShaderDefi) -> [VArray a] -> m ()

Use shader definitions for rendering.

isShaderCompiled :: (Farbe m, AttrType a b) => (b -> ShaderDefi) -> m Bool

module Graphics.Farbe.Vec

data Expr (e :: k) (a :: k1)

User-facing type expression.

data V

data F

fragCoord :: V4 (Expr F Float)

napier :: Fractional a => a

ln :: Floating a => a -> a

modf :: forall {k} (e :: k). Expr e Float -> Expr e Float -> Expr e Float

equot :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

erem :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

ediv :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

emod :: forall {k} (e :: k). Expr e Int32 -> Expr e Int32 -> Expr e Int32

transfer :: Transfer a b => a -> ShaderM b

use :: Use a e r => a -> r

Make shared variables for shaders

makeVarF :: (Farbe m, MonadIO m) => Float -> m (Var Float)

makeVarI :: (Farbe m, MonadIO m) => Int32 -> m (Var Int32)

makeVarB :: (Farbe m, MonadIO m) => Bool -> m (Var Bool)

makeVarV2F :: (Farbe m, MonadIO m) => V2 Float -> m (Var (V2 Float))

makeVarV2I :: (Farbe m, MonadIO m) => V2 Int32 -> m (Var (V2 Int32))

makeVarV2B :: (Farbe m, MonadIO m) => V2 Bool -> m (Var (V2 Bool))

makeVarV3F :: (Farbe m, MonadIO m) => V3 Float -> m (Var (V3 Float))

makeVarV3I :: (Farbe m, MonadIO m) => V3 Int32 -> m (Var (V3 Int32))

makeVarV3B :: (Farbe m, MonadIO m) => V3 Bool -> m (Var (V3 Bool))

makeVarV4F :: (Farbe m, MonadIO m) => V4 Float -> m (Var (V4 Float))

makeVarV4I :: (Farbe m, MonadIO m) => V4 Int32 -> m (Var (V4 Int32))

makeVarV4B :: (Farbe m, MonadIO m) => V4 Bool -> m (Var (V4 Bool))

makeVarM2 :: (Farbe m, MonadIO m) => V2 (V2 Float) -> m (Var (V2 (V2 Float)))

makeVarM3 :: (Farbe m, MonadIO m) => V3 (V3 Float) -> m (Var (V3 (V3 Float)))

makeVarM4 :: (Farbe m, MonadIO m) => V4 (V4 Float) -> m (Var (V4 (V4 Float)))

makeVarT :: (Farbe m, MonadIO m) => Texture -> m (Var Texture)

texture :: forall {k} (e :: k). Expr e Texture -> V2 (Expr e Float) -> V4 (Expr e Float)

data Texture

loadTexture :: (MonadIO m, HandTex m) => IO (TextureFormat, V2 GLsizei, Ptr a) -> m Texture

Vertex array

newtype VArray (a :: k)

Constructors

VArray

Fields

unVArray :: MVar (VArrayF a)

newVArray :: (HandVBO m, Storable a, Foldable f) => f a -> m (VArray a)

data Var a

Constructors

Var

Fields

varExpr :: ExprI

varMVar :: MVar a

swapVar :: MonadIO m => Var a -> a -> m a

class Storable a => AttrType a b | a -> b, b -> a

Minimal complete definition

setAttribute

Rendering control

drawOver :: MonadIO m => m a1 -> m a2 -> m ()

drawTexture :: (Monad m, Farbe m, MonadWindow m) => m (m () -> m Texture)

drawDepth :: (Monad m, Farbe m, MonadWindow m) => m (m () -> m Texture)

drawInto :: MonadIO m => m a1 -> m a2 -> m ()

Configuration options

modifyConfig :: Farbe m => (Config -> Config) -> m ()

data Config

Constructors

Config

Fields

debugMode :: Bool

devDebugMode :: Bool

workTime :: Double

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3

Miscellaneous

data FarbeT (m :: Type -> Type) a

class MonadIO m => Farbe (m :: Type -> Type)

Minimal complete definition

stateFarbe

runFarbeT' :: MonadIO m => FarbeT m a -> m a

Orphan instances

(Farbe m, Monad m) => Farbe (WindowT m)

Instance details

Methods

stateFarbe :: (FarbeState -> (a, FarbeState)) -> WindowT m a

getsFarbe :: (FarbeState -> a) -> WindowT m a

getFarbe :: WindowT m FarbeState

putFarbe :: FarbeState -> WindowT m ()

modifyFarbe :: (FarbeState -> FarbeState) -> WindowT m ()