Graphics.Farbe.Shader

Documentation

type ShaderM = DeferT' Shdr

type ShaderDefi = ShaderM (V4 (Expr V Float), V4 (Expr F Float))

compileShader :: (Farbe m, Farbe f, AttrType a b) => (b -> ShaderDefi) -> m (f Bool)

isShaderCompiled' :: MonadIO m => ShaderId -> m Bool

addShader :: (Farbe m, ShaderEnv m) => GLenum -> BuildShaderT m a -> m a

isShaderCompiled :: (Farbe m, AttrType a b) => (b -> ShaderDefi) -> m Bool

getExpr :: (Farbe m, AttrType a b) => (b -> ShaderDefi) -> m (V4 (Expr V Float), V4 (Expr F Float))

lookupShader :: (Farbe m, Farbe f, AttrType a b) => (b -> ShaderDefi) -> m (Maybe (f Bool))

shader :: (Farbe m, AttrType a b) => (b -> ShaderDefi) -> [VArray a] -> m ()

toCStatements :: [(String, ExprS)] -> String

toCExpr :: ExprS -> String

class Transfer a b | a -> b, b -> a where

Transfer values from vertex shader to fragment shader. Floating point numbers will be interpolated among its triangle space. Integers are taken from the first point of the triangle.

Methods

transfer :: a -> ShaderM b

transfer1 :: GLtype a => Expr V a -> DeferT' Shdr (Expr F a)

newtype DeferT n (m :: Type -> Type) a

DeferT, simple monad to defer monadic operations.

Constructors

DeferT

Fields

unDefer :: StateT (Seq n) m a

type DeferT' (m :: Type -> Type) = DeferT (m ()) m

runDeferT :: Monad m => DeferT n m a -> m (a, [n])

runDeferT' :: Monad m => DeferT' m a -> m a

runDeferT'' :: Monad m => DeferT n m a -> m (a, [n])

class Monad m => Defer n (m :: Type -> Type) | m -> n where

Methods

defer :: n -> m ()